kernel: huge refactoring
- Rewrote whole VMM
- Added userspace processes/threads support
- Implemented stack-trace
- Implemented syscalls
- Rewrote half of task switching code
- Rewrote int.s code
- Implemented thread sleeping
- Replaced multiboot.h to more correct one
- Added ELF32 format support
- Fixed some bugs including:
	- Wrong IDT loading
	- Very fast PIT (which was broke tasking sometimes)
	- Wrong memory mapping, because entire VMM was piece of bug
- Commented code for FPU/MMX/SIMD save/restore
- Not fully implemented code for processes/threads destroying
- Bochs config
- Rewrote some parts of utils.sh
- GPF plug

+++ src/sys/pic.c
@@ -45,8 +45,8 @@ void pic_sendEOI_slave() {
 }
 
 void pic_sendEOI(uint32_t isr) {
-	if (isr >= pic_master_offset && isr <= pic_slave_offset + 7) {
-		if (isr >= pic_slave_offset)
+	if ((uint32_t)isr >= pic_master_offset && isr <= (uint32_t)pic_slave_offset + 7) {
+		if ((uint32_t)isr >= pic_slave_offset)
 			pic_sendEOI_slave();
 
 		pic_sendEOI_master();
